Relocate keypairs.py from datastore to crypto
