+++ nucypher/crypto/powers.py
@@ -26,8 +26,8 @@ from nucypher.blockchain.eth.decorators import validate_checksum_address
 from nucypher.blockchain.eth.interfaces import BlockchainInterface, BlockchainInterfaceFactory
 from nucypher.blockchain.eth.signers.software import Web3Signer
 from nucypher.blockchain.eth.signers.base import Signer
-from nucypher.datastore import keypairs
-from nucypher.datastore.keypairs import DecryptingKeypair, SigningKeypair
+from nucypher.crypto import keypairs
+from nucypher.crypto.keypairs import DecryptingKeypair, SigningKeypair
 
 
 class PowerUpError(TypeError):
