+++ nucypher/network/server.py
@@ -23,7 +23,6 @@ from bytestring_splitter import BytestringSplitter
 from constant_sorrow import constants
 from constant_sorrow.constants import FLEET_STATES_MATCH, NO_BLOCKCHAIN_CONNECTION, NO_KNOWN_NODES
 from flask import Flask, Response, jsonify, request
-from hendrix.experience import crosstown_traffic
 from jinja2 import Template, TemplateError
 from typing import Tuple, Set
 from umbral.keys import UmbralPublicKey
@@ -33,12 +32,12 @@ from web3.exceptions import TimeExhausted
 import nucypher
 from nucypher.config.constants import MAX_UPLOAD_CONTENT_LENGTH
 from nucypher.config.storages import ForgetfulNodeStorage
+from nucypher.crypto.keypairs import HostingKeypair
 from nucypher.crypto.kits import UmbralMessageKit
 from nucypher.crypto.powers import KeyPairBasedPower, PowerUpError
 from nucypher.crypto.signing import InvalidSignature
 from nucypher.crypto.utils import canonical_address_from_umbral_key
 from nucypher.datastore.datastore import Datastore, RecordNotFound, DatastoreTransactionError
-from nucypher.datastore.keypairs import HostingKeypair
 from nucypher.datastore.models import PolicyArrangement, Workorder
 from nucypher.network import LEARNING_LOOP_VERSION
 from nucypher.network.exceptions import NodeSeemsToBeDown
